requirements.txt
langchain
langchain-community
openai
chromadb
python-dotenv
streamlit
tiktoken
